properties utils tests
